fix: fixed MANY errors in k8s files

+++ services/posts/PostService/Migrations/20221023185401_posts_init.up.sql
@@ -1,6 +1,6 @@
 ﻿CREATE TABLE IF NOT EXISTS Posts
 (
-    id int PRIMARY KEY,
+    id BIGSERIAL PRIMARY KEY,
     header VARCHAR(256) NOT NULL,
     body VARCHAR(1024) NOT NULL,
     author_id VARCHAR(256) NOT NULL
